requirements.txt
pandas==2.2.2
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
fuzzywuzzy==0.18.0
python-levenshtein==0.25.1
numpy==1.26.4